+++ src/zoid/buttons/util.js
@@ -102,7 +102,7 @@ export function createVenmoExperiment() : ?Experiment {
             return createExperiment('enable_venmo_android', 90);
         }
     } else {
-        return createExperiment('enable_venmo_desktop', 100);
+        return createExperiment('enable_venmo_desktop', 90);
     }
 }